requirements.txt
# https://pypi.org/
pandas==2.2.2
numpy==2.0.0
matplotlib==3.9.1
scikit-learn==1.5.1
joblib==1.4.2
supabase==2.5.1
python-dotenv==1.0.1 